requirements.txt
PyQt5>=5.15.0
python-docx>=0.8.11
num2words>=0.5.10
fpdf2>=2.5.0
msal>=1.20.0
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.1.0
arabic-reshaper>=2.1.3
python-bidi>=0.4.2
beautifulsoup4>=4.11.0